Add -clean=exit
With -clean=exit, goswarm automatically cleans up instances on exit
(either due to signal or test completion).

This is helpful when you don't need to do anything with the gomote after
finding a failure. In particular, it ensures that only exactly the
goswarm created are the ones that get destroyed, preventing you from
accidentally interfering with gomotes used for other purposes.